refactor(c5c3-operator): split reconcile_korc.go into per-area files
reconcile_korc.go held three sub-reconcilers and eight distinct
responsibilities in one 1.4k-line file. Move each cohesive group into
its own file, following the file-per-sub-reconciler convention already
used elsewhere in the package. Pure code motion — no symbol is renamed
and no behaviour changes.

- reconcile_korc.go (~440 lines): AC mint/re-mint state machine plus
  the shared conditionFailer/crdMissing helpers.
- reconcile_admincredential.go: reconcileAdminCredential and
  pushSecretReady.
- reconcile_catalog.go: reconcileCatalog and the catalog naming/URL
  helpers.
- korc_secrets.go: owned-Secret ensure/seed/regenerate helpers and
  the admin-password read/hash helpers.
- korc_cloudsyaml.go: pure clouds.yaml builders and their key const.
- korc_eso.go: PushSecret/ExternalSecret builders and the remote-key
  helper.
- korc_imports.go: ensureKORCAdminImports and the admin user/domain
  refs.

pushSecretReady stays local rather than moving to
internal/common/secrets: the keystone operator has no equivalent
predicate today, so the move is deferred to its second consumer.
Build, gofmt, go vet, golangci-lint, unit and integration tests pass.

AI-assisted: Claude Code
On-behalf-of: @SAP christian.berendt@sap.com
