+++ operators/c5c3/internal/controller/reconcile_admincredential.go
@@ -1,0 +1,195 @@
+// SPDX-FileCopyrightText: Copyright 2026 SAP SE or an SAP affiliate company
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package controller
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+
+	esov1alpha1 "github.com/external-secrets/external-secrets/apis/externalsecrets/v1alpha1"
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/log"
+
+	"github.com/c5c3/forge/internal/common/conditions"
+	"github.com/c5c3/forge/internal/common/secrets"
+	c5c3v1alpha1 "github.com/c5c3/forge/operators/c5c3/api/v1alpha1"
+)
+
+// reconcileAdminCredential commits the minted application credential into an
+// operator-owned Secret and mirrors it to OpenBao, driving the
+// AdminCredentialReady condition.
+//
+// It is GATED on KORCReady: until reconcileKORC reports the AC minted there is
+// nothing to push. It is additionally gated on the K-ORC clouds.yaml
+// ExternalSecret being Ready in the ControlPlane's OWN namespace
+// (childNamespace(cp)/<CloudCredentialsRef.SecretName>) — co-located with the
+// K-ORC CRs per C1 — so the credential is never published before K-ORC can
+// authenticate with its admin cloud.
+func (r *ControlPlaneReconciler) reconcileAdminCredential(ctx context.Context, cp *c5c3v1alpha1.ControlPlane) (ctrl.Result, error) {
+	logger := log.FromContext(ctx)
+	fail := conditionFailer(cp, conditionTypeAdminCredentialReady)
+
+	// Gate on KORCReady.
+	if !conditions.AllTrue(cp.Status.Conditions, conditionTypeKORCReady) {
+		logger.Info("KORC not ready, deferring admin credential push")
+		fail("WaitingForKORC", "KORCReady is not True; admin credential push deferred")
+		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
+	}
+
+	// Gate on the OpenBao-backed ClusterSecretStore so an ESO/OpenBao outage
+	// surfaces as AdminCredentialReady=False promptly. The clouds.yaml
+	// ExternalSecret read below would eventually requeue on its own, but ESO only
+	// re-syncs at the refreshInterval (default 1h); the ClusterSecretStore watch
+	// wakes the ControlPlane the moment ESO flips the store condition, so the
+	// credential condition does not stay stale-True through a short outage (#476).
+	storeReady, err := secrets.IsClusterSecretStoreReady(ctx, r.Client, openBaoClusterStoreName)
+	if err != nil {
+		return ctrl.Result{}, err
+	}
+	if !storeReady {
+		logger.Info("ClusterSecretStore not ready, deferring admin credential push")
+		conditions.SetCondition(&cp.Status.Conditions, metav1.Condition{
+			Type:               conditionTypeAdminCredentialReady,
+			Status:             metav1.ConditionFalse,
+			ObservedGeneration: cp.Generation,
+			Reason:             "SecretStoreNotReady",
+			Message: fmt.Sprintf("ClusterSecretStore %q is not ready; upstream secret backend unreachable",
+				openBaoClusterStoreName),
+		})
+		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
+	}
+
+	// Gate on the K-ORC clouds.yaml ExternalSecret being Ready. It MUST materialise
+	// in the SAME namespace as the K-ORC resource CRs (childNamespace) because
+	// K-ORC resolves CloudCredentialsRef in the resource's own namespace (C1). The
+	// Secret name follows the spec's CloudCredentialsRef.SecretName — the exact
+	// value reconcileKORC sets on the AC CR — defaulted to korcCloudsYamlSecretName
+	// by the webhook (the fallback below covers a webhook-bypass edge case).
+	cloudsYamlName := cp.Spec.KORC.AdminCredential.CloudCredentialsRef.SecretName
+	if cloudsYamlName == "" {
+		cloudsYamlName = korcCloudsYamlSecretName
+	}
+	ready, err := secrets.WaitForExternalSecret(ctx, r.Client,
+		types.NamespacedName{Namespace: childNamespace(cp), Name: cloudsYamlName})
+	if err != nil {
+		fail("CloudsYamlError", fmt.Sprintf("checking k-orc clouds.yaml ExternalSecret: %v", err))
+		return ctrl.Result{}, err
+	}
+	if !ready {
+		logger.Info("k-orc clouds.yaml ExternalSecret not ready, requeuing")
+		fail("WaitingForCloudsYaml", "k-orc clouds.yaml ExternalSecret is not yet Ready")
+		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
+	}
+
+	// Assemble the application-credential clouds.yaml into the operator-owned Secret
+	// so the PushSecret mirrors it to OpenBao (and ESO re-materialises it as the
+	// admin clouds.yaml, replacing the password-based bootstrap seed). K-ORC does
+	// NOT write this — it only consumed Secret.Data["value"] to mint — so we build
+	// it from the minted credential id (AC status, surfaced on cp.Status) and the
+	// generated secret value. The "value" key is preserved untouched.
+	acID := ""
+	if cp.Status.AdminApplicationCredential != nil {
+		acID = cp.Status.AdminApplicationCredential.ID
+	}
+	if acID == "" {
+		logger.Info("admin application credential id not yet reported, deferring credential assembly")
+		fail("WaitingForCredentialID", "minted application credential id is not yet reported by K-ORC")
+		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
+	}
+
+	// The operator-owned Secret (with its generated "value" and owner reference)
+	// was created by ensureAppCredentialSecret during reconcileKORC, which KORCReady
+	// — gated above — guarantees has run and K-ORC has read the "value" to mint.
+	// Get it directly rather than CreateOrUpdate: a NotFound here is an invariant
+	// violation (the minted secret vanished), NOT a create opportunity — minting a
+	// fresh "value" would not match the credential Keystone already issued, so we
+	// requeue and let reconcileKORC re-establish the Secret instead of writing a
+	// brand-new empty one that would immediately fail the "value" check.
+	secretKey := types.NamespacedName{
+		Name:      adminAppCredentialSecretName(cp),
+		Namespace: childNamespace(cp),
+	}
+	secret := &corev1.Secret{}
+	if err := r.Get(ctx, secretKey, secret); err != nil {
+		if apierrors.IsNotFound(err) {
+			logger.Info("app-credential secret not found, deferring credential assembly", "secret", secretKey.Name)
+			fail("WaitingForAppCredentialSecret", fmt.Sprintf("app-credential secret %q does not exist yet", secretKey.Name))
+			return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
+		}
+		fail("SecretError", fmt.Sprintf("reading admin app-credential secret %q: %v", secretKey.Name, err))
+		return ctrl.Result{}, err
+	}
+
+	value := secret.Data[appCredSecretValueKey]
+	if len(value) == 0 {
+		logger.Info("app-credential secret has no value yet, deferring credential assembly", "secret", secretKey.Name)
+		fail("WaitingForAppCredentialSecret", fmt.Sprintf("app-credential secret %q has no %q key (mint not complete?)",
+			secret.Name, appCredSecretValueKey))
+		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
+	}
+
+	// Persist the assembled clouds.yaml under appCredCloudsYAMLKey, leaving the
+	// "value" key untouched. Skip the write when it already matches so repeated
+	// reconciles do not churn the Secret (and wake ESO to re-push).
+	cloudsYAML := []byte(buildAppCredCloudsYAML(cp, acID, string(value)))
+	if !bytes.Equal(secret.Data[appCredCloudsYAMLKey], cloudsYAML) {
+		secret.Data[appCredCloudsYAMLKey] = cloudsYAML
+		if err := r.Update(ctx, secret); err != nil {
+			fail("SecretError", fmt.Sprintf("assembling admin app-credential clouds.yaml: %v", err))
+			return ctrl.Result{}, err
+		}
+	}
+
+	// CLOBBER-SAFE PushSecret: EnsurePushSecret is idempotent — it only Updates
+	// the PushSecret when its desired Spec differs from the stored one
+	// (apiequality.Semantic.DeepEqual guard inside EnsurePushSecret). Repeated
+	// reconciles therefore do not churn the PushSecret, so ESO is not woken to
+	// re-push an unchanged credential.
+	ps := adminAppCredentialPushSecret(cp)
+	if err := secrets.EnsurePushSecret(ctx, r.Client, r.Scheme, cp, ps); err != nil {
+		fail("PushSecretError", fmt.Sprintf("ensuring admin app-credential PushSecret: %v", err))
+		return ctrl.Result{}, err
+	}
+
+	// Gate AdminCredentialReady on the PushSecret actually syncing to OpenBao — not
+	// merely on the CR existing. Otherwise a backend permission failure (e.g. the
+	// ESO role missing the push-app-credentials policy) yields a false-positive
+	// Ready while OpenBao still serves the password-based bootstrap clouds.yaml.
+	pushed := &esov1alpha1.PushSecret{}
+	if err := r.Get(ctx, types.NamespacedName{Name: ps.Name, Namespace: ps.Namespace}, pushed); err != nil {
+		fail("PushSecretError", fmt.Sprintf("reading admin app-credential PushSecret: %v", err))
+		return ctrl.Result{}, err
+	}
+	if !pushSecretReady(pushed) {
+		logger.Info("admin app-credential PushSecret not yet synced, requeuing")
+		fail("WaitingForPushSecret", "admin app-credential PushSecret has not synced to OpenBao yet")
+		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
+	}
+
+	conditions.SetCondition(&cp.Status.Conditions, metav1.Condition{
+		Type:               conditionTypeAdminCredentialReady,
+		Status:             metav1.ConditionTrue,
+		ObservedGeneration: cp.Generation,
+		Reason:             "AdminCredentialReady",
+		Message:            "Admin application credential committed and mirrored to OpenBao",
+	})
+	return ctrl.Result{}, nil
+}
+
+// pushSecretReady reports whether an ESO PushSecret has synced to its backend
+// (its "Ready" condition is True).
+func pushSecretReady(ps *esov1alpha1.PushSecret) bool {
+	for _, c := range ps.Status.Conditions {
+		if c.Type == esov1alpha1.PushSecretReady && c.Status == corev1.ConditionTrue {
+			return true
+		}
+	}
+	return false
+}
