+++ operators/c5c3/internal/controller/reconcile_korc.go
@@ -5,25 +5,16 @@
 package controller
 
 import (
-	"bytes"
 	"context"
-	"crypto/rand"
-	"crypto/sha256"
-	"encoding/base64"
-	"encoding/hex"
 	"fmt"
 	"time"
 
-	esov1 "github.com/external-secrets/external-secrets/apis/externalsecrets/v1"
-	esov1alpha1 "github.com/external-secrets/external-secrets/apis/externalsecrets/v1alpha1"
 	orcv1alpha1 "github.com/k-orc/openstack-resource-controller/v2/api/v1alpha1"
-	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/utils/ptr"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
@@ -32,285 +23,21 @@ import (
 	c5c3v1alpha1 "github.com/c5c3/forge/operators/c5c3/api/v1alpha1"
 )
 
-// openBaoClusterStoreName mirrors the keystone operator's constant
-// (operators/keystone/internal/controller/reconcile_secrets.go): the cluster-
-// scoped ESO SecretStore that fronts OpenBao. Defined locally so the c5c3
-// PushSecrets target the same backend without importing the keystone package
-const openBaoClusterStoreName = "openbao-cluster-store"
-
 // adminAppCredentialNameSuffix is appended to the ControlPlane name to derive
 // the deterministic, collision-free name of the owned K-ORC ApplicationCredential
 // CR, mirroring the keystoneNameSuffix discipline so a single
 // namespace can host the admin AC of multiple ControlPlanes.
 const adminAppCredentialNameSuffix = "-admin-app-credential" //nolint:gosec // G101 false positive: name suffix, not a credential.
 
-// adminAppCredentialSecretSuffix names the operator-owned Secret that K-ORC
-// writes the minted application credential into (Resource.SecretRef). It is the
-// push source for the OpenBao PushSecret.
-const adminAppCredentialSecretSuffix = "-admin-app-credential" //nolint:gosec // G101 false positive: Secret name suffix, not a credential.
-
-// adminAppCredentialRemoteKeyFor returns the per-ControlPlane OpenBao path the
-// minted admin application credential is mirrored to. The
-// path is scoped by both the ControlPlane's Namespace and Name so two
-// ControlPlanes never clobber each other's admin credential on the cluster-
-// global OpenBao backend. The matching read consumer (the per-CR k-orc
-// clouds.yaml ExternalSecret) targets the same key.
-func adminAppCredentialRemoteKeyFor(cp *c5c3v1alpha1.ControlPlane) string {
-	return fmt.Sprintf("openstack/keystone/%s/%s/admin/app-credential", cp.Namespace, cp.Name)
-}
-
 // adminPasswordHashAnnotation stamps the SHA-256 of the admin password the
 // application credential was last minted against onto the owned AC CR. Mirrors the hash+annotation pattern in the keystone operator's
 // password-rotation reconciler. A mismatch on a later pass drives a re-mint.
 const adminPasswordHashAnnotation = "forge.c5c3.io/admin-password-hash" //nolint:gosec // G101 false positive: annotation key, not a credential.
 
-// korcCloudsYamlSecretName is the conventional name of the admin clouds.yaml
-// Secret (and its ExternalSecret) K-ORC reads its admin credentials from. It
-// matches DefaultCloudCredentialsSecretName, the value the defaulting webhook
-// applies to spec.korc.adminCredential.cloudCredentialsRef.secretName, and is
-// used only as the fallback when a CR somehow reaches the reconciler without the
-// webhook having defaulted that field.
-//
-// C1 (co-location): the ExternalSecret materialises this Secret into the
-// SAME namespace as the K-ORC ApplicationCredential/Service/Endpoint CRs the
-// operator projects — the ControlPlane's own namespace (childNamespace(cp)) —
-// because K-ORC resolves CloudCredentialsRef in the resource's own namespace
-// (vendored api/v1alpha1/credentials_ref.go GetCloudCredentialsRef returns the
-// resource's Namespace). The AdminCredentialReady gate therefore waits on the
-// ExternalSecret in childNamespace(cp), NOT a fixed orc-system one, so the minted
-// AC is never pushed before K-ORC can actually authenticate.
-const korcCloudsYamlSecretName = "k-orc-clouds-yaml" //nolint:gosec // G101 false positive: Secret name, not a credential.
-
-// adminPasswordCloudSecretSuffix names the operator-owned Secret holding a
-// PASSWORD-based clouds.yaml that always tracks the current admin password. The
-// admin ApplicationCredential mints against THIS secret (not k-orc-clouds-yaml),
-// which breaks the self-referential bootstrap deadlock: k-orc-clouds-yaml is the
-// minted app-credential itself, so deleting the AC to re-mint would invalidate
-// the very clouds.yaml needed to re-authenticate. A restricted application
-// credential also cannot mint a new application credential — only a
-// password-authenticated session can — so password auth is required for the
-// delete+recreate re-mint to work at all.
-const adminPasswordCloudSecretSuffix = "-admin-password-cloud" //nolint:gosec // G101 false positive: Secret name suffix, not a credential.
-
 // adminAppCredentialName returns the deterministic name of the owned K-ORC
 // ApplicationCredential CR for the given ControlPlane.
 func adminAppCredentialName(cp *c5c3v1alpha1.ControlPlane) string {
 	return cp.Name + adminAppCredentialNameSuffix
-}
-
-// adminPasswordCloudSecretName returns the name of the operator-owned Secret that
-// holds the password-based clouds.yaml the admin ApplicationCredential mints with.
-func adminPasswordCloudSecretName(cp *c5c3v1alpha1.ControlPlane) string {
-	return cp.Name + adminPasswordCloudSecretSuffix
-}
-
-// adminAppCredentialSecretName returns the name of the operator-owned Secret that
-// holds the application-credential secret K-ORC mints with (key "value") and,
-// after minting, the assembled app-credential clouds.yaml pushed to OpenBao.
-func adminAppCredentialSecretName(cp *c5c3v1alpha1.ControlPlane) string {
-	return cp.Name + adminAppCredentialSecretSuffix
-}
-
-// appCredSecretValueKey is the Secret data key K-ORC reads the application
-// credential's secret from (the actuator reads Secret.Data["value"]).
-const appCredSecretValueKey = "value"
-
-// appCredCloudsYAMLKey is the Secret data key the assembled app-credential
-// clouds.yaml is stored under; the PushSecret mirrors it to OpenBao and the
-// k-orc-clouds-yaml ExternalSecret reads it back as the "clouds.yaml" property.
-const appCredCloudsYAMLKey = "clouds.yaml"
-
-// ensureOwnedSecret create-or-updates an operator-owned corev1.Secret named
-// `name` in childNamespace(cp), with cp set as the controller owner reference.
-// The Secret's Data map is guaranteed non-nil before `mutate` runs, so callers
-// only set the keys they own; `mutate` may return an error to abort the write
-// (e.g. when generating a random value fails). It centralises the four
-// near-identical owned-Secret CreateOrUpdate wrappers (ensureAppCredentialSecret,
-// ensureAdminPasswordCloud, seedBootstrapCloudsYAML and
-// regenerateAppCredentialSecretValue); each keeps its own error wrapping so the
-// failure context stays specific.
-func (r *ControlPlaneReconciler) ensureOwnedSecret(
-	ctx context.Context, cp *c5c3v1alpha1.ControlPlane, name string, mutate func(*corev1.Secret) error,
-) error {
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: childNamespace(cp),
-		},
-	}
-	_, err := controllerutil.CreateOrUpdate(ctx, r.Client, secret, func() error {
-		if secret.Data == nil {
-			secret.Data = map[string][]byte{}
-		}
-		if err := mutate(secret); err != nil {
-			return err
-		}
-		return controllerutil.SetControllerReference(cp, secret, r.Scheme)
-	})
-	return err
-}
-
-// ensureAppCredentialSecret ensures the operator-owned Secret that K-ORC reads the
-// application-credential secret from exists with a generated "value". K-ORC's
-// managed ApplicationCredential reads Secret.Data["value"] and creates the AC in
-// Keystone with it, so this MUST exist before the AC is reconciled. The value is
-// generated once and preserved across reconciles — regenerating it would force a
-// re-mint and invalidate the stored clouds.yaml.
-func (r *ControlPlaneReconciler) ensureAppCredentialSecret(ctx context.Context, cp *c5c3v1alpha1.ControlPlane) error {
-	if err := r.ensureOwnedSecret(ctx, cp, adminAppCredentialSecretName(cp), func(secret *corev1.Secret) error {
-		if len(secret.Data[appCredSecretValueKey]) == 0 {
-			v, gerr := generateAppCredSecretValue()
-			if gerr != nil {
-				return gerr
-			}
-			secret.Data[appCredSecretValueKey] = []byte(v)
-		}
-		return nil
-	}); err != nil {
-		return fmt.Errorf("ensuring app-credential secret %q: %w", adminAppCredentialSecretName(cp), err)
-	}
-	return nil
-}
-
-// generateAppCredSecretValue returns a 256-bit, URL-safe random string used as the
-// application credential's secret.
-func generateAppCredSecretValue() (string, error) {
-	b := make([]byte, 32)
-	if _, err := rand.Read(b); err != nil {
-		return "", fmt.Errorf("generating application-credential secret: %w", err)
-	}
-	return base64.RawURLEncoding.EncodeToString(b), nil
-}
-
-// buildAppCredCloudsYAML assembles the application-credential clouds.yaml the
-// control plane authenticates K-ORC with after minting: the credential id comes
-// from the minted AC, the secret from the generated "value", and the auth_url from
-// the projected Keystone Service (keystoneEndpointURL).
-//
-// CRITICAL (endpoint_type: internal): gophercloud only uses the auth_url to obtain
-// a token; for every subsequent API call it resolves the endpoint from the returned
-// service catalog, picking the interface set here. K-ORC runs IN-CLUSTER, so it
-// must use the "internal" (cluster-DNS) identity endpoint. Once the ControlPlane
-// exposes Keystone via the shared Gateway the catalog's "public" identity endpoint
-// becomes the external host (e.g. https://keystone.<host>.nip.io:8443/v3), which
-// from inside a pod is unreachable — so "public" makes every list/get fail. Worse,
-// K-ORC swallows that failure (osclients ListDomains does `_ = pager.EachPage(...)`)
-// and reports it as an EMPTY import, so the admin Domain/User imports hang forever
-// on "Waiting for OpenStack resource to be created externally".
-//
-// The key MUST be "endpoint_type", NOT "interface": K-ORC's scope builder copies
-// only clientconfig.Cloud.EndpointType (the `endpoint_type` key) into the client
-// options and drops Cloud.Interface (the `interface` key) — see vendored
-// internal/scope/provider.go NewProviderClient. An "interface:" value is therefore
-// ignored and the endpoint defaults to "public". The auth_url already points at the
-// in-cluster Service for the same reason (keystoneEndpointURL, never the external
-// endpoint).
-func buildAppCredCloudsYAML(cp *c5c3v1alpha1.ControlPlane, acID, secret string) string {
-	region := cp.Spec.Region
-	if region == "" {
-		region = "RegionOne"
-	}
-	return fmt.Sprintf(`clouds:
-  admin:
-    auth:
-      auth_url: %q
-      application_credential_id: %q
-      application_credential_secret: %q
-    auth_type: v3applicationcredential
-    region_name: %q
-    endpoint_type: internal
-    identity_api_version: 3
-`, keystoneEndpointURL(cp), acID, secret, region)
-}
-
-// ensureAdminPasswordCloud ensures the operator-owned Secret holding the
-// PASSWORD-based clouds.yaml the admin ApplicationCredential mints with exists and
-// always tracks the CURRENT admin password. Unlike ensureAppCredentialSecret's
-// "value" (generated once and preserved), this clouds.yaml is rebuilt from the
-// live admin password on every pass, so a password rotation flows through to it —
-// the credential K-ORC uses to re-authenticate and revoke/re-mint never goes
-// stale. CreateOrUpdate only writes when the rendered clouds.yaml differs, so a
-// steady-state pass does not churn the Secret (and does not wake any consumer).
-//
-// The Secret lives in childNamespace(cp) — the same namespace as the K-ORC CRs —
-// because K-ORC resolves CloudCredentialsRef in the resource's own namespace (C1).
-func (r *ControlPlaneReconciler) ensureAdminPasswordCloud(ctx context.Context, cp *c5c3v1alpha1.ControlPlane, password string) error {
-	if err := r.ensureOwnedSecret(ctx, cp, adminPasswordCloudSecretName(cp), func(secret *corev1.Secret) error {
-		secret.Data[appCredCloudsYAMLKey] = []byte(buildPasswordCloudsYAML(cp, password))
-		return nil
-	}); err != nil {
-		return fmt.Errorf("ensuring admin password-cloud secret %q: %w", adminPasswordCloudSecretName(cp), err)
-	}
-	return nil
-}
-
-// buildPasswordCloudsYAML assembles the password-based clouds.yaml the admin
-// ApplicationCredential authenticates with to mint (and, on re-mint, revoke) the
-// Keystone credential. It mirrors the bootstrap seed
-// (deploy/openbao/bootstrap/write-bootstrap-secrets.sh) so the in-cluster and
-// operator-owned credentials are byte-compatible: the cloud key matches the
-// CloudCredentialsRef.CloudName, auth_url is the in-cluster Keystone Service
-// (keystoneEndpointURL — never the external endpoint), and endpoint_type is
-// "internal" (the key MUST be "endpoint_type", not "interface"; see
-// buildAppCredCloudsYAML for the full rationale).
-func buildPasswordCloudsYAML(cp *c5c3v1alpha1.ControlPlane, password string) string {
-	region := cp.Spec.Region
-	if region == "" {
-		region = "RegionOne"
-	}
-	cloudName := cp.Spec.KORC.AdminCredential.CloudCredentialsRef.CloudName
-	if cloudName == "" {
-		cloudName = korcAdminUsername
-	}
-	return fmt.Sprintf(`clouds:
-  %s:
-    auth:
-      auth_url: %q
-      username: %q
-      password: %q
-      project_name: %q
-      user_domain_name: %q
-      project_domain_name: %q
-    region_name: %q
-    endpoint_type: internal
-    identity_api_version: 3
-`, cloudName, keystoneEndpointURL(cp), korcAdminUsername, password,
-		korcAdminUsername, korcAdminDomainName, korcAdminDomainName, region)
-}
-
-// seedBootstrapCloudsYAML writes the PASSWORD-based clouds.yaml into the
-// {cp.Name}-admin-app-credential Secret's clouds.yaml key, but ONLY when that key
-// is empty (write-if-empty). It breaks the AdminCredentialReady chicken-and-egg
-// deadlock on a fresh cluster the per-CR OpenBao path the
-// k-orc-clouds-yaml ExternalSecret reads is empty until something pushes to it, so
-// the operator seeds a password-based document here that lets K-ORC's admin
-// imports authenticate before the application credential is ever minted —
-// previously this was seeded by deploy/openbao/bootstrap/write-bootstrap-secrets.sh.
-//
-// Write-if-empty is the idempotency guard once
-// reconcileAdminCredential fills the key with the minted credential-based
-// clouds.yaml (buildAppCredCloudsYAML) the seed becomes a no-op and never clobbers
-// the minted document. On a re-mint, regenerateAppCredentialSecretValue deletes
-// the key, so the next reconcileKORC pass re-seeds the password version and bridges
-// the re-authentication gap. The "value" key (owned by ensureAppCredentialSecret)
-// is never touched.
-func (r *ControlPlaneReconciler) seedBootstrapCloudsYAML(ctx context.Context, cp *c5c3v1alpha1.ControlPlane, password string) error {
-	if err := r.ensureOwnedSecret(ctx, cp, adminAppCredentialSecretName(cp), func(secret *corev1.Secret) error {
-		// Write-if-empty: never overwrite a minted credential-based clouds.yaml.
-		if len(secret.Data[appCredCloudsYAMLKey]) == 0 {
-			secret.Data[appCredCloudsYAMLKey] = []byte(buildPasswordCloudsYAML(cp, password))
-		}
-		return nil
-	}); err != nil {
-		return fmt.Errorf("seeding bootstrap clouds.yaml into secret %q: %w", adminAppCredentialSecretName(cp), err)
-	}
-	return nil
-}
-
-// adminAppCredentialPushSecretName returns the PushSecret name backing up the
-// minted application credential to OpenBao.
-func adminAppCredentialPushSecretName(cp *c5c3v1alpha1.ControlPlane) string {
-	return cp.Name + "-admin-app-credential-backup"
 }
 
 // conditionFailer returns a closure bound to cp and condType that stamps a
@@ -605,94 +332,6 @@ func (r *ControlPlaneReconciler) remintAdminApplicationCredential(
 	return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
 }
 
-// regenerateAppCredentialSecretValue overwrites the app-credential secret "value"
-// with a fresh random secret and drops any stale assembled clouds.yaml. The new
-// "value" makes the recreated AC mint a NEW Keystone credential; dropping the
-// clouds.yaml forces reconcileAdminCredential to rebuild it from the fresh id+value
-// rather than keep serving the just-revoked credential.
-func (r *ControlPlaneReconciler) regenerateAppCredentialSecretValue(ctx context.Context, cp *c5c3v1alpha1.ControlPlane) error {
-	if err := r.ensureOwnedSecret(ctx, cp, adminAppCredentialSecretName(cp), func(secret *corev1.Secret) error {
-		v, gerr := generateAppCredSecretValue()
-		if gerr != nil {
-			return gerr
-		}
-		secret.Data[appCredSecretValueKey] = []byte(v)
-		delete(secret.Data, appCredCloudsYAMLKey)
-		return nil
-	}); err != nil {
-		return fmt.Errorf("regenerating app-credential secret value: %w", err)
-	}
-	return nil
-}
-
-// adminUserRef returns the Kubernetes metadata.name of the imported K-ORC User
-// CR the admin application credential is associated with. AdminCredentialSpec has
-// no UserRef field, but K-ORC's ApplicationCredentialResourceSpec.UserRef is
-// REQUIRED, so we derive a deterministic name scoped by cp.Name (mirroring
-// adminDomainRef) — this way two ControlPlanes in one namespace produce DISTINCT
-// User objects rather than colliding on a shared name. The
-// inner OpenStack username the import resolves to is still "admin": it is set
-// independently via Spec.Import.Filter.Name = OpenStackName(korcAdminUsername) in
-// ensureKORCAdminImports. The matching User CR is provisioned there as an
-// unmanaged import, so the reference always resolves.
-func adminUserRef(cp *c5c3v1alpha1.ControlPlane) string {
-	return fmt.Sprintf("%s-user-admin", cp.Name)
-}
-
-// korcAdminUsername / korcAdminDomainName identify the OpenStack admin user and
-// its domain that the Keystone bootstrap creates; the c5c3-operator imports them
-// into K-ORC (unmanaged) rather than creating them.
-const (
-	korcAdminUsername   = "admin"
-	korcAdminDomainName = "Default"
-)
-
-// adminDomainRef is the deterministic name of the K-ORC Domain CR the admin User
-// import is scoped to.
-func adminDomainRef(cp *c5c3v1alpha1.ControlPlane) string {
-	return cp.Name + "-domain-default"
-}
-
-// ensureKORCAdminImports ensures the K-ORC Domain and User that the admin
-// ApplicationCredential's UserRef depends on exist as UNMANAGED imports. The
-// Keystone bootstrap creates the real admin user (in the Default domain); K-ORC
-// must import — not create — it, otherwise the ApplicationCredential blocks on
-// "Waiting for User/admin to be created". Both CRs are owned by the ControlPlane
-// for GC and reuse the admin clouds.yaml credentials.
-func (r *ControlPlaneReconciler) ensureKORCAdminImports(ctx context.Context, cp *c5c3v1alpha1.ControlPlane, credRef orcv1alpha1.CloudCredentialsReference) error {
-	domain := &orcv1alpha1.Domain{
-		ObjectMeta: metav1.ObjectMeta{Name: adminDomainRef(cp), Namespace: childNamespace(cp)},
-	}
-	if _, err := controllerutil.CreateOrUpdate(ctx, r.Client, domain, func() error {
-		domain.Spec.ManagementPolicy = orcv1alpha1.ManagementPolicyUnmanaged
-		domain.Spec.CloudCredentialsRef = credRef
-		domain.Spec.Import = &orcv1alpha1.DomainImport{
-			Filter: &orcv1alpha1.DomainFilter{Name: ptr.To(orcv1alpha1.KeystoneName(korcAdminDomainName))},
-		}
-		return controllerutil.SetControllerReference(cp, domain, r.Scheme)
-	}); err != nil {
-		return fmt.Errorf("admin Domain import %q: %w", domain.Name, err)
-	}
-
-	user := &orcv1alpha1.User{
-		ObjectMeta: metav1.ObjectMeta{Name: adminUserRef(cp), Namespace: childNamespace(cp)},
-	}
-	if _, err := controllerutil.CreateOrUpdate(ctx, r.Client, user, func() error {
-		user.Spec.ManagementPolicy = orcv1alpha1.ManagementPolicyUnmanaged
-		user.Spec.CloudCredentialsRef = credRef
-		user.Spec.Import = &orcv1alpha1.UserImport{
-			Filter: &orcv1alpha1.UserFilter{
-				Name:      ptr.To(orcv1alpha1.OpenStackName(korcAdminUsername)),
-				DomainRef: ptr.To(orcv1alpha1.KubernetesNameRef(adminDomainRef(cp))),
-			},
-		}
-		return controllerutil.SetControllerReference(cp, user, r.Scheme)
-	}); err != nil {
-		return fmt.Errorf("admin User import %q: %w", user.Name, err)
-	}
-	return nil
-}
-
 // projectAccessRules maps our AccessRule{Service,Method,Path} list onto K-ORC's
 // ApplicationCredentialAccessRule list. K-ORC models the service as a serviceRef
 // (a reference to an ORC Service CR named after the service type) and the method
@@ -734,50 +373,6 @@ func projectAccessRules(rules []c5c3v1alpha1.AccessRule) []orcv1alpha1.Applicati
 	return out
 }
 
-// computeAdminPasswordHash reads the admin password from the ControlPlane's
-// configured PasswordSecretRef and returns its SHA-256 as a lowercase hex string
-// The data key defaults to "password" when the SecretRef.Key
-// is unset, matching the keystone admin-password Secret convention.
-//
-// DECISION (hash-helper extraction): the hash derivation lives here as a
-// package-level function so BOTH the ControlPlane reconciler (which reads the
-// cleartext via readAdminPassword and hashes it via hashAdminPassword inline) and
-// the CredentialRotation reconciler compute the SAME hash without duplicating the
-// SHA-256 logic.
-func computeAdminPasswordHash(ctx context.Context, c client.Client, cp *c5c3v1alpha1.ControlPlane) (string, error) {
-	pw, err := readAdminPassword(ctx, c, cp)
-	if err != nil {
-		return "", err
-	}
-	return hashAdminPassword(pw), nil
-}
-
-// readAdminPassword reads the cleartext admin password from the EFFECTIVE
-// admin-password Secret (data key defaults to "password"). The effective ref
-// is the operator-owned per-ControlPlane Secret
-// adminPasswordSecretName(cp) in managed mode and the user-supplied
-// cp.Spec.KORC.AdminCredential.PasswordSecretRef in brownfield mode — see
-// effectiveAdminPasswordSecretRef. reconcileKORC needs the cleartext — not just
-// its hash — to render the password-based clouds.yaml the admin
-// ApplicationCredential mints with, so the read is factored out here and the
-// hash derived from it via hashAdminPassword.
-func readAdminPassword(ctx context.Context, c client.Client, cp *c5c3v1alpha1.ControlPlane) (string, error) {
-	ref := effectiveAdminPasswordSecretRef(cp)
-	key := ref.Key
-	if key == "" {
-		key = "password"
-	}
-	return secrets.GetSecretValue(ctx, c,
-		types.NamespacedName{Namespace: cp.Namespace, Name: ref.Name}, key)
-}
-
-// hashAdminPassword returns the SHA-256 of the admin password as a lowercase hex
-// string — the value stamped onto the AC CR's adminPasswordHashAnnotation.
-func hashAdminPassword(pw string) string {
-	sum := sha256.Sum256([]byte(pw))
-	return hex.EncodeToString(sum[:])
-}
-
 // updateAdminApplicationCredentialStatus reflects the observed AC CR into
 // cp.Status.AdminApplicationCredential. LastRotation is
 // (re-)stamped to now whenever the recorded credential ID changes (initial mint
@@ -812,392 +407,4 @@ func (r *ControlPlaneReconciler) updateAdminApplicationCredentialStatus(
 		status.LastRotation = prev.LastRotation
 	}
 	cp.Status.AdminApplicationCredential = status
-}
-
-// reconcileAdminCredential commits the minted application credential into an
-// operator-owned Secret and mirrors it to OpenBao, driving the
-// AdminCredentialReady condition.
-//
-// It is GATED on KORCReady: until reconcileKORC reports the AC minted there is
-// nothing to push. It is additionally gated on the K-ORC clouds.yaml
-// ExternalSecret being Ready in the ControlPlane's OWN namespace
-// (childNamespace(cp)/<CloudCredentialsRef.SecretName>) — co-located with the
-// K-ORC CRs per C1 — so the credential is never published before K-ORC can
-// authenticate with its admin cloud.
-func (r *ControlPlaneReconciler) reconcileAdminCredential(ctx context.Context, cp *c5c3v1alpha1.ControlPlane) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-	fail := conditionFailer(cp, conditionTypeAdminCredentialReady)
-
-	// Gate on KORCReady.
-	if !conditions.AllTrue(cp.Status.Conditions, conditionTypeKORCReady) {
-		logger.Info("KORC not ready, deferring admin credential push")
-		fail("WaitingForKORC", "KORCReady is not True; admin credential push deferred")
-		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
-	}
-
-	// Gate on the OpenBao-backed ClusterSecretStore so an ESO/OpenBao outage
-	// surfaces as AdminCredentialReady=False promptly. The clouds.yaml
-	// ExternalSecret read below would eventually requeue on its own, but ESO only
-	// re-syncs at the refreshInterval (default 1h); the ClusterSecretStore watch
-	// wakes the ControlPlane the moment ESO flips the store condition, so the
-	// credential condition does not stay stale-True through a short outage (#476).
-	storeReady, err := secrets.IsClusterSecretStoreReady(ctx, r.Client, openBaoClusterStoreName)
-	if err != nil {
-		return ctrl.Result{}, err
-	}
-	if !storeReady {
-		logger.Info("ClusterSecretStore not ready, deferring admin credential push")
-		conditions.SetCondition(&cp.Status.Conditions, metav1.Condition{
-			Type:               conditionTypeAdminCredentialReady,
-			Status:             metav1.ConditionFalse,
-			ObservedGeneration: cp.Generation,
-			Reason:             "SecretStoreNotReady",
-			Message: fmt.Sprintf("ClusterSecretStore %q is not ready; upstream secret backend unreachable",
-				openBaoClusterStoreName),
-		})
-		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
-	}
-
-	// Gate on the K-ORC clouds.yaml ExternalSecret being Ready. It MUST materialise
-	// in the SAME namespace as the K-ORC resource CRs (childNamespace) because
-	// K-ORC resolves CloudCredentialsRef in the resource's own namespace (C1). The
-	// Secret name follows the spec's CloudCredentialsRef.SecretName — the exact
-	// value reconcileKORC sets on the AC CR — defaulted to korcCloudsYamlSecretName
-	// by the webhook (the fallback below covers a webhook-bypass edge case).
-	cloudsYamlName := cp.Spec.KORC.AdminCredential.CloudCredentialsRef.SecretName
-	if cloudsYamlName == "" {
-		cloudsYamlName = korcCloudsYamlSecretName
-	}
-	ready, err := secrets.WaitForExternalSecret(ctx, r.Client,
-		types.NamespacedName{Namespace: childNamespace(cp), Name: cloudsYamlName})
-	if err != nil {
-		fail("CloudsYamlError", fmt.Sprintf("checking k-orc clouds.yaml ExternalSecret: %v", err))
-		return ctrl.Result{}, err
-	}
-	if !ready {
-		logger.Info("k-orc clouds.yaml ExternalSecret not ready, requeuing")
-		fail("WaitingForCloudsYaml", "k-orc clouds.yaml ExternalSecret is not yet Ready")
-		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
-	}
-
-	// Assemble the application-credential clouds.yaml into the operator-owned Secret
-	// so the PushSecret mirrors it to OpenBao (and ESO re-materialises it as the
-	// admin clouds.yaml, replacing the password-based bootstrap seed). K-ORC does
-	// NOT write this — it only consumed Secret.Data["value"] to mint — so we build
-	// it from the minted credential id (AC status, surfaced on cp.Status) and the
-	// generated secret value. The "value" key is preserved untouched.
-	acID := ""
-	if cp.Status.AdminApplicationCredential != nil {
-		acID = cp.Status.AdminApplicationCredential.ID
-	}
-	if acID == "" {
-		logger.Info("admin application credential id not yet reported, deferring credential assembly")
-		fail("WaitingForCredentialID", "minted application credential id is not yet reported by K-ORC")
-		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
-	}
-
-	// The operator-owned Secret (with its generated "value" and owner reference)
-	// was created by ensureAppCredentialSecret during reconcileKORC, which KORCReady
-	// — gated above — guarantees has run and K-ORC has read the "value" to mint.
-	// Get it directly rather than CreateOrUpdate: a NotFound here is an invariant
-	// violation (the minted secret vanished), NOT a create opportunity — minting a
-	// fresh "value" would not match the credential Keystone already issued, so we
-	// requeue and let reconcileKORC re-establish the Secret instead of writing a
-	// brand-new empty one that would immediately fail the "value" check.
-	secretKey := types.NamespacedName{
-		Name:      adminAppCredentialSecretName(cp),
-		Namespace: childNamespace(cp),
-	}
-	secret := &corev1.Secret{}
-	if err := r.Get(ctx, secretKey, secret); err != nil {
-		if apierrors.IsNotFound(err) {
-			logger.Info("app-credential secret not found, deferring credential assembly", "secret", secretKey.Name)
-			fail("WaitingForAppCredentialSecret", fmt.Sprintf("app-credential secret %q does not exist yet", secretKey.Name))
-			return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
-		}
-		fail("SecretError", fmt.Sprintf("reading admin app-credential secret %q: %v", secretKey.Name, err))
-		return ctrl.Result{}, err
-	}
-
-	value := secret.Data[appCredSecretValueKey]
-	if len(value) == 0 {
-		logger.Info("app-credential secret has no value yet, deferring credential assembly", "secret", secretKey.Name)
-		fail("WaitingForAppCredentialSecret", fmt.Sprintf("app-credential secret %q has no %q key (mint not complete?)",
-			secret.Name, appCredSecretValueKey))
-		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
-	}
-
-	// Persist the assembled clouds.yaml under appCredCloudsYAMLKey, leaving the
-	// "value" key untouched. Skip the write when it already matches so repeated
-	// reconciles do not churn the Secret (and wake ESO to re-push).
-	cloudsYAML := []byte(buildAppCredCloudsYAML(cp, acID, string(value)))
-	if !bytes.Equal(secret.Data[appCredCloudsYAMLKey], cloudsYAML) {
-		secret.Data[appCredCloudsYAMLKey] = cloudsYAML
-		if err := r.Update(ctx, secret); err != nil {
-			fail("SecretError", fmt.Sprintf("assembling admin app-credential clouds.yaml: %v", err))
-			return ctrl.Result{}, err
-		}
-	}
-
-	// CLOBBER-SAFE PushSecret: EnsurePushSecret is idempotent — it only Updates
-	// the PushSecret when its desired Spec differs from the stored one
-	// (apiequality.Semantic.DeepEqual guard inside EnsurePushSecret). Repeated
-	// reconciles therefore do not churn the PushSecret, so ESO is not woken to
-	// re-push an unchanged credential.
-	ps := adminAppCredentialPushSecret(cp)
-	if err := secrets.EnsurePushSecret(ctx, r.Client, r.Scheme, cp, ps); err != nil {
-		fail("PushSecretError", fmt.Sprintf("ensuring admin app-credential PushSecret: %v", err))
-		return ctrl.Result{}, err
-	}
-
-	// Gate AdminCredentialReady on the PushSecret actually syncing to OpenBao — not
-	// merely on the CR existing. Otherwise a backend permission failure (e.g. the
-	// ESO role missing the push-app-credentials policy) yields a false-positive
-	// Ready while OpenBao still serves the password-based bootstrap clouds.yaml.
-	pushed := &esov1alpha1.PushSecret{}
-	if err := r.Get(ctx, types.NamespacedName{Name: ps.Name, Namespace: ps.Namespace}, pushed); err != nil {
-		fail("PushSecretError", fmt.Sprintf("reading admin app-credential PushSecret: %v", err))
-		return ctrl.Result{}, err
-	}
-	if !pushSecretReady(pushed) {
-		logger.Info("admin app-credential PushSecret not yet synced, requeuing")
-		fail("WaitingForPushSecret", "admin app-credential PushSecret has not synced to OpenBao yet")
-		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
-	}
-
-	conditions.SetCondition(&cp.Status.Conditions, metav1.Condition{
-		Type:               conditionTypeAdminCredentialReady,
-		Status:             metav1.ConditionTrue,
-		ObservedGeneration: cp.Generation,
-		Reason:             "AdminCredentialReady",
-		Message:            "Admin application credential committed and mirrored to OpenBao",
-	})
-	return ctrl.Result{}, nil
-}
-
-// pushSecretReady reports whether an ESO PushSecret has synced to its backend
-// (its "Ready" condition is True).
-func pushSecretReady(ps *esov1alpha1.PushSecret) bool {
-	for _, c := range ps.Status.Conditions {
-		if c.Type == esov1alpha1.PushSecretReady && c.Status == corev1.ConditionTrue {
-			return true
-		}
-	}
-	return false
-}
-
-// adminAppCredentialPushSecret builds the PushSecret that mirrors the minted
-// admin application-credential Secret to OpenBao at the per-ControlPlane path
-// openstack/keystone/{cp.Namespace}/{cp.Name}/admin/app-credential
-// scoping the credential so two ControlPlanes never clobber
-// each other's admin credential on the cluster-global OpenBao backend.
-//
-// DECISION (DeletionPolicy): None — the admin application credential is a shared
-// bootstrap secret other consumers may depend on; deleting the PushSecret (e.g.
-// on ControlPlane teardown) leaves the last-pushed credential intact in OpenBao
-// so a fresh control plane is not locked out mid-rotation. Reviewer: please verify.
-func adminAppCredentialPushSecret(cp *c5c3v1alpha1.ControlPlane) *esov1alpha1.PushSecret {
-	return &esov1alpha1.PushSecret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      adminAppCredentialPushSecretName(cp),
-			Namespace: childNamespace(cp),
-		},
-		Spec: esov1alpha1.PushSecretSpec{
-			DeletionPolicy: esov1alpha1.PushSecretDeletionPolicyNone,
-			SecretStoreRefs: []esov1alpha1.PushSecretStoreRef{{
-				Kind: "ClusterSecretStore",
-				Name: openBaoClusterStoreName,
-			}},
-			Selector: esov1alpha1.PushSecretSelector{
-				Secret: &esov1alpha1.PushSecretSecret{
-					Name: adminAppCredentialSecretName(cp),
-				},
-			},
-			Data: []esov1alpha1.PushSecretData{{
-				Match: esov1alpha1.PushSecretMatch{
-					RemoteRef: esov1alpha1.PushSecretRemoteRef{
-						RemoteKey: adminAppCredentialRemoteKeyFor(cp),
-					},
-				},
-			}},
-		},
-	}
-}
-
-// ensureKORCCloudsYAMLExternalSecret create-or-updates the per-ControlPlane,
-// operator-owned ExternalSecret that materialises the admin clouds.yaml Secret
-// K-ORC authenticates with, replacing the retired static single-identity manifest
-// It is created in childNamespace(cp) — co-located with the
-// K-ORC resource CRs because K-ORC resolves CloudCredentialsRef in the resource's
-// own namespace (C1) — and reads the per-CR OpenBao path
-// adminAppCredentialRemoteKeyFor(cp), so an arbitrarily named ControlPlane resolves
-// to the correct key with no manifest edit.
-//
-// The Secret name follows the spec's CloudCredentialsRef.SecretName (defaulted to
-// korcCloudsYamlSecretName by the webhook; the fallback covers a webhook-bypass
-// edge case). CreationPolicy Owner makes ESO own the materialised Secret, and the
-// ExternalSecret itself is owner-referenced to the ControlPlane for GC. The
-// ExternalSecret type is esov1 (NOT esov1alpha1 as the PushSecret above is).
-func (r *ControlPlaneReconciler) ensureKORCCloudsYAMLExternalSecret(ctx context.Context, cp *c5c3v1alpha1.ControlPlane) error {
-	name := cp.Spec.KORC.AdminCredential.CloudCredentialsRef.SecretName
-	if name == "" {
-		name = korcCloudsYamlSecretName
-	}
-	es := &esov1.ExternalSecret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: childNamespace(cp),
-		},
-	}
-	if _, err := controllerutil.CreateOrUpdate(ctx, r.Client, es, func() error {
-		es.Spec.RefreshInterval = &metav1.Duration{Duration: time.Hour}
-		es.Spec.SecretStoreRef = esov1.SecretStoreRef{
-			Kind: "ClusterSecretStore",
-			Name: openBaoClusterStoreName,
-		}
-		es.Spec.Target = esov1.ExternalSecretTarget{
-			Name:           name,
-			CreationPolicy: esov1.CreatePolicyOwner,
-		}
-		es.Spec.Data = []esov1.ExternalSecretData{{
-			SecretKey: appCredCloudsYAMLKey,
-			RemoteRef: esov1.ExternalSecretDataRemoteRef{
-				Key:      adminAppCredentialRemoteKeyFor(cp),
-				Property: appCredCloudsYAMLKey,
-			},
-		}}
-		return controllerutil.SetControllerReference(cp, es, r.Scheme)
-	}); err != nil {
-		return fmt.Errorf("ensuring k-orc clouds.yaml ExternalSecret %q: %w", es.Name, err)
-	}
-	return nil
-}
-
-// reconcileCatalog registers the OpenStack service catalog entries for Keystone
-// (an identity Service plus its public Endpoint) as OWNED K-ORC CRs and drives
-// the CatalogReady condition.
-//
-// It is GATED on AdminCredentialReady: the admin credential must be available
-// before K-ORC can register catalog entries. Both child CRs are create-or-updated
-// idempotently; CatalogReady (and cp.Status.CatalogReady) flip True once both are
-// registered. K-ORC is a hard CRD dependency (see reconcileKORC), so a missing
-// Service/Endpoint CRD never reaches here — no-match errors fall through to the
-// generic error returns below (#476).
-func (r *ControlPlaneReconciler) reconcileCatalog(ctx context.Context, cp *c5c3v1alpha1.ControlPlane) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-	fail := conditionFailer(cp, conditionTypeCatalogReady)
-
-	// Gate on AdminCredentialReady.
-	if !conditions.AllTrue(cp.Status.Conditions, conditionTypeAdminCredentialReady) {
-		logger.Info("AdminCredential not ready, deferring catalog registration")
-		fail("WaitingForAdminCredential", "AdminCredentialReady is not True; catalog registration deferred")
-		return ctrl.Result{RequeueAfter: korcRequeueAfter}, nil
-	}
-
-	secretName := cp.Spec.KORC.AdminCredential.CloudCredentialsRef.SecretName
-	// Fall back to the conventional name when SecretName is empty, matching
-	// reconcileAdminCredential and ensureKORCCloudsYAMLExternalSecret so a
-	// webhook-bypass CR resolves to the same clouds.yaml Secret name everywhere
-	// (#476). Without this the catalog Service/Endpoint would reference an empty
-	// CloudCredentialsRef.SecretName.
-	if secretName == "" {
-		secretName = korcCloudsYamlSecretName
-	}
-	cloudName := cp.Spec.KORC.AdminCredential.CloudCredentialsRef.CloudName
-	credRef := orcv1alpha1.CloudCredentialsReference{SecretName: secretName, CloudName: cloudName}
-
-	// 1. Identity (Keystone) Service.
-	service := &orcv1alpha1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      keystoneServiceName(cp),
-			Namespace: childNamespace(cp),
-		},
-	}
-	if _, err := controllerutil.CreateOrUpdate(ctx, r.Client, service, func() error {
-		service.Spec.ManagementPolicy = orcv1alpha1.ManagementPolicyManaged
-		service.Spec.CloudCredentialsRef = credRef
-		if service.Spec.Resource == nil {
-			service.Spec.Resource = &orcv1alpha1.ServiceResourceSpec{}
-		}
-		service.Spec.Resource.Type = "identity"
-		service.Spec.Resource.Name = ptr.To(orcv1alpha1.OpenStackName("keystone"))
-		service.Spec.Resource.Enabled = ptr.To(true)
-		return controllerutil.SetControllerReference(cp, service, r.Scheme)
-	}); err != nil {
-		fail("ServiceError", fmt.Sprintf("create-or-update identity Service: %v", err))
-		return ctrl.Result{}, err
-	}
-
-	// 2. Public Endpoint for the Keystone API.
-	endpoint := &orcv1alpha1.Endpoint{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      keystoneEndpointName(cp),
-			Namespace: childNamespace(cp),
-		},
-	}
-	if _, err := controllerutil.CreateOrUpdate(ctx, r.Client, endpoint, func() error {
-		endpoint.Spec.ManagementPolicy = orcv1alpha1.ManagementPolicyManaged
-		endpoint.Spec.CloudCredentialsRef = credRef
-		if endpoint.Spec.Resource == nil {
-			endpoint.Spec.Resource = &orcv1alpha1.EndpointResourceSpec{}
-		}
-		endpoint.Spec.Resource.Interface = "public"
-		// DECISION (Endpoint URL): K-ORC's EndpointResourceSpec.URL is REQUIRED.
-		// When the ControlPlane exposes Keystone externally (a gateway or explicit
-		// publicEndpoint is set) we register that public URL so the catalog matches
-		// what Keystone's own bootstrap advertises; otherwise we fall back to the
-		// in-cluster Keystone Service URL derived from the PROJECTED Keystone
-		// Service — keystoneName(cp) = "{cp.Name}-keystone" in the ControlPlane
-		// namespace — which is the Service the keystone-operator actually exposes.
-		endpoint.Spec.Resource.URL = keystoneCatalogURL(cp)
-		endpoint.Spec.Resource.ServiceRef = orcv1alpha1.KubernetesNameRef(keystoneServiceName(cp))
-		endpoint.Spec.Resource.Enabled = ptr.To(true)
-		return controllerutil.SetControllerReference(cp, endpoint, r.Scheme)
-	}); err != nil {
-		fail("EndpointError", fmt.Sprintf("create-or-update identity Endpoint: %v", err))
-		return ctrl.Result{}, err
-	}
-
-	cp.Status.CatalogReady = true
-	conditions.SetCondition(&cp.Status.Conditions, metav1.Condition{
-		Type:               conditionTypeCatalogReady,
-		Status:             metav1.ConditionTrue,
-		ObservedGeneration: cp.Generation,
-		Reason:             "CatalogRegistered",
-		Message:            "Keystone identity Service and Endpoint are registered",
-	})
-	return ctrl.Result{}, nil
-}
-
-// keystoneServiceName / keystoneEndpointName return the deterministic names of
-// the owned K-ORC Service/Endpoint CRs registering the identity catalog entry.
-func keystoneServiceName(cp *c5c3v1alpha1.ControlPlane) string {
-	return cp.Name + "-identity-service"
-}
-
-func keystoneEndpointName(cp *c5c3v1alpha1.ControlPlane) string {
-	return cp.Name + "-identity-endpoint"
-}
-
-// keystoneEndpointURL derives the in-cluster Keystone identity URL from the
-// projected Keystone Service — keystoneName(cp) = "{cp.Name}-keystone" — in the
-// ControlPlane namespace (see DECISION on Endpoint URL in reconcileCatalog). It
-// must NOT hard-code "keystone": the keystone-operator names the Service after
-// the projected Keystone CR, so a fixed name would not resolve. This is the URL
-// K-ORC authenticates against (the seeded clouds.yaml auth_url): K-ORC runs
-// in-cluster, so it must always use the Service DNS, never the external endpoint.
-func keystoneEndpointURL(cp *c5c3v1alpha1.ControlPlane) string {
-	return fmt.Sprintf("http://%s.%s.svc:5000/v3", keystoneName(cp), childNamespace(cp))
-}
-
-// keystoneCatalogURL returns the URL registered for the K-ORC identity catalog
-// Endpoint. It prefers the externally routable publicEndpoint (keystonePublicEndpoint)
-// so the catalog matches what Keystone's own bootstrap advertises when exposed
-// via a Gateway; absent external exposure it falls back to the in-cluster
-// Service URL (keystoneEndpointURL).
-func keystoneCatalogURL(cp *c5c3v1alpha1.ControlPlane) string {
-	if pe := keystonePublicEndpoint(cp.Spec.Services.Keystone); pe != "" {
-		return pe
-	}
-	return keystoneEndpointURL(cp)
 }
